+++ src/interface/meshblock_data.hpp
@@ -21,6 +21,7 @@
 #include <vector>
 
 #include "interface/data_collection.hpp"
+#include "interface/sparse_pack_base.hpp"
 #include "interface/variable.hpp"
 #include "interface/variable_pack.hpp"
 #include "mesh/domain.hpp"
@@ -243,6 +244,8 @@ class MeshBlockData {
   /// included if sparse_ids is not empty and contains the sparse id of the sparse
   /// variable
 
+  SparsePackCache &GetSparsePackCache() { return sparse_pack_cache_; }
+
   /// Pack variables and fluxes by separate variables and fluxes names
   const VariableFluxPack<T> &
   PackVariablesAndFluxes(const std::vector<std::string> &var_names,
@@ -463,8 +466,8 @@ class MeshBlockData {
                                  "', but no such sparse variable exists");
 
     auto var = GetCellVarPtr(label);
-    PARTHENON_REQUIRE_THROWS(var->IsSparse(),
-                             "Tried to deallocate non-sparse variable " + label);
+    // PARTHENON_REQUIRE_THROWS(var->IsSparse(),
+    //                         "Tried to deallocate non-sparse variable " + label);
 
     if (var->IsAllocated()) {
       var->Deallocate();
@@ -484,6 +487,7 @@ class MeshBlockData {
   MapToVariablePack<T> varPackMap_;
   MapToVariablePack<T> coarseVarPackMap_; // cache for varpacks over coarse arrays
   MapToVariableFluxPack<T> varFluxPackMap_;
+  SparsePackCache sparse_pack_cache_;
 
   // These functions have private scope and are visible only to MeshData
   const VariableFluxPack<T> &